+++ src/search.cpp
@@ -687,19 +687,19 @@ Value Worker::search(
             if (SEE::value(captured) > SEE::value(PieceType::Pawn)) {
                 if (non_pawn_material < 0) {
                     non_pawn_material =
-                      static_cast<i32>(pos.ipiece_count(Color::White, PieceType::Queen)
+                      (pos.ipiece_count(Color::White, PieceType::Queen)
                                        + pos.ipiece_count(Color::Black, PieceType::Queen))
                       * SEE::value(PieceType::Queen);
                     non_pawn_material +=
-                      static_cast<i32>(pos.ipiece_count(Color::White, PieceType::Rook)
+                      (pos.ipiece_count(Color::White, PieceType::Rook)
                                        + pos.ipiece_count(Color::Black, PieceType::Rook))
                       * SEE::value(PieceType::Rook);
                     non_pawn_material +=
-                      static_cast<i32>(pos.ipiece_count(Color::White, PieceType::Bishop)
+                      (pos.ipiece_count(Color::White, PieceType::Bishop)
                                        + pos.ipiece_count(Color::Black, PieceType::Bishop))
                       * SEE::value(PieceType::Bishop);
                     non_pawn_material +=
-                      static_cast<i32>(pos.ipiece_count(Color::White, PieceType::Knight)
+                      (pos.ipiece_count(Color::White, PieceType::Knight)
                                        + pos.ipiece_count(Color::Black, PieceType::Knight))
                       * SEE::value(PieceType::Knight);
                 }
